Lots o stuff.  This version is built with the introspecive lenda event class
Has new and imporved corections reading system.  returns vectors in one line
as opposed to one value at a time.  Looks for the first occurance of the string
given with a _1 after it. So CorMan->GetVec("goe1") looks for goe1_1 in file
contiunes until it sees a different key.

In addition to the normal correction definion lines like Event->SetPositionCorreection(..)
It has the Define correction statements.  The old explict corrections have been commeneted out in
lenda event to improve speed.

Also has a timing feature.  After the first 20k events it meassures the time it took and gives an
estimate untill done